backend/predator-ai/requirements.txt
# Core
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
websockets==12.0
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.0

# Blockchain
web3==6.11.1
eth-account==0.9.0
moralis==1.12.0
solana==0.30.2
ton-client-py==1.0.0

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.2
torch==2.1.0
transformers==4.35.0
tensorflow==2.13.0
stable-baselines3==2.1.0
gym==0.26.2
networkx==3.1
dgl==1.1.2

# Social Media
tweepy==4.14.0
telethon==1.34.0
discord.py==2.3.2

# Database
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.0

# API & Monitoring
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
python-telegram-bot==20.6

# Utilities
loguru==0.7.2
tenacity==8.2.3
python-json-logger==2.0.7